남현수 sprint6

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -8,19 +8,19 @@ import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.ChannelType;
 import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.ReadStatus;
+import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -32,18 +32,26 @@ public class BasicChannelService implements ChannelService {
   private final MessageRepository messageRepository;
 
   @Override
+  @Transactional
   public Channel createPublicChannel(PublicChannelCreateRequest request) {
     // Public은 생성시 기존 로직을 유지.
-    Channel channel = new Channel(ChannelType.PUBLIC, request.name(), request.description());
+    Channel channel = Channel.builder()
+        .type(ChannelType.PUBLIC)
+        .name(request.name())
+        .description(request.description())
+        .build();
     channelRepository.save(channel);
 
     return channel;
   }
 
   @Override
+  @Transactional
   public Channel createPrivateChannel(PrivateChannelCreateRequest request) {
     // Private 채널 생성 로직 (name, description 은 생략함.)
-    Channel channel = new Channel(ChannelType.PRIVATE, null, null);
+    Channel channel = Channel.builder()
+        .type(ChannelType.PRIVATE)
+        .build();
 
     // 1. Channel에 참여하려는 User의 정보를 DTO를 통해서 받아 user 별 ReadStatus 정보를 생성한다.
     for (UUID userId : request.participantIds()) {
@@ -52,7 +60,7 @@ public class BasicChannelService implements ChannelService {
       userRepository.findById(userId)
           .orElseThrow(() -> new NoSuchElementException("존재하지 않는 회원입니다."));
 
-      ReadStatus readStatus = new ReadStatus(userId, channel.getId(), Instant.MIN);
+      ReadStatus readStatus = ReadStatus.builder().build();
       readStatusRepository.save(readStatus);
     }
 
@@ -62,6 +70,7 @@ public class BasicChannelService implements ChannelService {
   }
 
   @Override
+  @Transactional
   public ChannelFindResponse findByChannelId(UUID channelId) {
 
     // 1. 호환성 확인
@@ -69,46 +78,48 @@ public class BasicChannelService implements ChannelService {
         .orElseThrow(() -> new NoSuchElementException("존재하지 않는 채널입니다."));
 
     // 2. 해당 채널의 가장 최근 메시지의 시간 정보를 포함.
-    Message latestMessageTime = messageRepository.latestMessageByChannelId(channelId)
+    Message latestMessage = messageRepository.findLatestMessage(channelId)
         .orElseThrow(() -> new IllegalArgumentException("일치하는 채널 아이디가 없습니다."));
     // 2-2. Private 채널인 경우 참여한 User의 Id 정보까지 포함시킴.
-    List<UUID> participantsIds = new ArrayList<>();
+    List<User> participantsIds = new ArrayList<>();
 
     if (channel.getType() == ChannelType.PRIVATE) {
       participantsIds = readStatusRepository.findAllByChannelId(channelId).stream()
-          .map(ReadStatus::getUserId).toList();
+          .map(ReadStatus::getUser)
+          .toList();
     }
 
     // DTO를 통해
     return ChannelFindResponse.builder().id(channelId).type(channel.getType())
         .name(channel.getName()).description(channel.getDescription())
-        .lastMessageAt(latestMessageTime.getCreatedAt()).participantIds(participantsIds).build();
+        .lastMessageAt(latestMessage.getCreatedAt()).participantIds(participantsIds).build();
   }
 
   @Override
+  @Transactional
   public List<ChannelFindResponse> findAllByUserId(UUID userId) {
     List<ChannelFindResponse> responses = new ArrayList<>();
 
     // 호환성 체크
     userRepository.findById(userId).orElseThrow(() -> new NoSuchElementException("존재하지 않는 회원입니다."));
     // 1. Private 채널은 User가 참여한 채널만 조회하기 위한 ChannelId를 List화 시킴.
-    List<UUID> subscribeChannelIds = readStatusRepository.findAllByUserId(userId).stream()
-        .map(ReadStatus::getChannelId).toList();
+    List<Channel> subscribeChannelIds = readStatusRepository.findAllByUserId(userId).stream()
+        .map(ReadStatus::getChannel).toList();
 
     // 2. Public의 Channel도 보여줘야 한다.
     for (Channel channel : channelRepository.findAll()) {
       // ChannelType이 Public 경우와 User가 참여한 채널일 경우는 List에 넣게 된다.
       if (channel.getType() == ChannelType.PUBLIC
           || channel.getType() == ChannelType.PRIVATE && subscribeChannelIds.contains(
-          channel.getId())) {
+          channel)) {
 
         // 3-1. 해당 채널의 가장 최근 메시지의 시간 정보를 포함시킴 없으면 null
-        Message latestMessageTime = messageRepository.latestMessageByChannelId(channel.getId())
+        Message latestMessageTime = messageRepository.findLatestMessage(channel.getId())
             .orElse(null);
 
         // 3-2. 그 채널의 참여한 모든 UserId를 불러옴.
-        List<UUID> participantsIds = readStatusRepository.findAllByChannelId(channel.getId())
-            .stream().map(ReadStatus::getUserId).toList();
+        List<User> participantsIds = readStatusRepository.findAllByChannelId(channel.getId())
+            .stream().map(ReadStatus::getUser).toList();
 
         // 4. DTO에 값 추가.
         responses.add(ChannelFindResponse.builder().id(channel.getId()).type(channel.getType())
@@ -122,6 +133,7 @@ public class BasicChannelService implements ChannelService {
   }
 
   @Override
+  @Transactional
   public Channel updateChannel(UUID channelId, ChannelUpdateRequest request) {
     // 1. 호환성 체크
     Channel channel = channelRepository.findById(channelId)
@@ -131,18 +143,18 @@ public class BasicChannelService implements ChannelService {
     if (channel.getType() == ChannelType.PRIVATE) {
       throw new IllegalArgumentException("PRIVATE 채널은 수정할 수 없습니다.");
     }
-
-    // DTO를 통한 update
-    channel.update(request.newName(), request.newDescription());
-
     return channelRepository.save(channel);
   }
 
   @Override
+  @Transactional
   public void deleteChannel(UUID channelId) {
 
     // 관련 도메인 같이 삭제 Channel, Message, ReadStatus
-    channelRepository.delete(channelId);
+    if (!channelRepository.existsById(channelId)) {
+      throw new NoSuchElementException("존재하지 않는 채널 입니다");
+    }
+    channelRepository.deleteById(channelId);
 
     messageRepository.deleteByChannelId(channelId);
